% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_convert.R
\name{expression_convert}
\alias{expression_convert}
\title{Convert New Expression}
\usage{
expression_convert(x)
}
\arguments{
\item{x}{A string}
}
\value{
An expression
}
\description{
Takes the new_expr and removes the for loop and add cbind
}
\examples{
expression_convert("log(eCount[i]) <- b0")
expression_convert("for(i in 1:nObs) {eCount[i] <- b0}")
expression_convert("for(i in 1:length(LogLength)) {eWeightLength[i] <- b0 + bDayte * Dayte[i]}")
expression_convert("for(i in 1:nObs) {eAnnual[i] <- bAnn[Ann[i]] + bSA[Site[i], Ann[i]]}")
}
