% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-vectorize.R
\name{expression_vectorize}
\alias{expression_vectorize}
\title{Convert New Expression}
\usage{
expression_vectorize(x)
}
\arguments{
\item{x}{An expression}
}
\value{
An expression
}
\description{
Takes an expression and removes the for loop and adds \code{cbind} for arrays.
}
\examples{
expression_vectorize(rlang::expr(for(i in 1:nObs) {eCount[i] <- b0}))
expression_vectorize(
  rlang::expr(
    for(i in 1:length(LogLength)) {eWeightLength[i] <- b0 + bDayte * Dayte[i]}
  )
)
expression_vectorize(
  rlang::expr(
    for(i in 1:nObs) {eAnnual[i] <- bAnn[Ann[i]] + bSA[Site[i], Ann[i]]}
  )
)
}
