% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcderive-package.R
\docType{package}
\name{mcmcderive-package}
\alias{mcmcderive}
\alias{mcmcderive-package}
\title{mcmcderive: Derive MCMC Parameters}
\description{
Generates derived parameters from Monte Carlo Markov Chain (MCMC) samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/mcmcderive}
  \item Report bugs at \url{https://github.com/poissonconsulting/mcmcderive/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
