% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-derive.R
\name{mcmc_derive}
\alias{mcmc_derive}
\alias{mcmc_derive.mcmcr}
\alias{mcmc_derive.mcmc_data}
\title{MCMC Derive}
\usage{
mcmc_derive(object, ...)

\method{mcmc_derive}{mcmcr}(object, expr, values = list(),
  monitor = ".*", parallel = FALSE, ...)

\method{mcmc_derive}{mcmc_data}(object, expr, values = list(),
  monitor = "prediction", parallel = FALSE, ...)
}
\arguments{
\item{object}{The MCMC object.}

\item{...}{Unused.}

\item{expr}{A string of the R expression to evaluate.}

\item{values}{A named list of additional values to evaluate in the R expression.}

\item{monitor}{A regular expression specifying the new variables to monitor.}

\item{parallel}{A flag indicating whether to derive samples in parallel using foreach backend.}
}
\value{
An \code{\link[mcmcr]{mcmcr}} object of the derived parameter(s).
}
\description{
Calculate derived parameters for an MCMC object.
}
\details{
In the case of an mcmc_data object, values includes the vectors 
in the associated data frame. 
The mcmcr parameter is named estimate (and this is the default variable to monitor).
}
\section{Methods (by class)}{
\itemize{
\item \code{mcmcr}: MCMC Derive for an mcmcr object

\item \code{mcmc_data}: MCMC Derive for an mcmc_data object
}}

\examples{
mcmc_derive(mcmcr::mcmcr_example, "prediction <- (alpha + beta) / sigma")
}
