% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcderive-package.R
\docType{package}
\name{mcmcderive-package}
\alias{mcmcderive}
\alias{mcmcderive-package}
\title{mcmcderive: Derive MCMC Parameters}
\description{
Generates derived parameter(s) from Monte Carlo
    Markov Chain (MCMC) samples using R code. This allows Bayesian models
    to be fitted without the inclusion of derived parameters which add
    unnecessary clutter and slow model fitting. For more information on
    MCMC samples see Brooks et al. (2011) <isbn:978-1-4200-7941-8>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/mcmcderive}
  \item Report bugs at \url{https://github.com/poissonconsulting/mcmcderive/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
